Fix: 댓글 조회 시 토큰이 없는 경우에도 정상 조회되도록 변경

+++ src/main/java/com/greedy/mokkoji/api/comment/dto/request/CommentCreateRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.Min;
 public record CommentCreateRequest(
         @Min(0)
         @Max(5)
-        Integer rate,
+        Double rate,
         String content
 ) {
 }
